Simplify MultiVector
