+++ src/linearAlgebra/MultiVector.t.cpp
@@ -20,288 +20,539 @@
  ******************************************************************************/
 
 /*
- * @author Sambit Das
+ * @author Sambit Das, Bikash Kanungo
  */
-
 #include <utils/Exceptions.h>
+#include <cmath>
 
 namespace dftefe
 {
   namespace linearAlgebra
   {
+    /**
+     * @brief Constructor for a \b serial MultiVector using size, numVectors and
+     * init value
+     **/
+    template <typename ValueType, dftefe::utils::MemorySpace memorySpace>
+    MultiVector<ValueType, memorySpace>::MultiVector(
+      const size_type               size,
+      const size_type               numVectors,
+      LinAlgOpContext<memorySpace> *linAlgOpContext,
+      const ValueType               initVal)
+    {
+      d_storage =
+        std::make_unique<typename MultiVector<ValueType, memorySpace>::Storage>(
+          size * numVectors, initVal);
+      d_linAlgOpContext = linAlgOpContext;
+      d_vectorAttributes =
+        VectorAttributes(VectorAttributes::Distribution::SERIAL);
+      d_globalSize       = size;
+      d_locallyOwnedSize = size;
+      d_ghostSize        = 0;
+      d_localSize        = d_locallyOwnedSize + d_ghostSize;
+      d_numVectors       = numVectors;
+      d_mpiPatternP2P =
+        std::make_shared<const utils::mpi::MPIPatternP2P<memorySpace>>(size);
+      d_mpiCommunicatorP2P = std::make_unique<
+        utils::mpi::MPICommunicatorP2P<ValueType, memorySpace>>(d_mpiPatternP2P,
+                                                                numVectors);
+    }
+
+    /**
+     * @brief Constructor for a \serial MultiVector with a predefined
+     * MultiVector::Storage (i.e., utils::MemoryStorage).
+     * This constructor transfers the ownership of the input Storage to the
+     * MultiVector. This is useful when one does not want to allocate new
+     * memory and instead use memory allocated in the MultiVector::Storage
+     * (i.e., utils::MemoryStorage).
+     */
+    template <typename ValueType, dftefe::utils::MemorySpace memorySpace>
+    MultiVector<ValueType, memorySpace>::MultiVector(
+      std::unique_ptr<typename MultiVector<ValueType, memorySpace>::Storage>
+                                    storage,
+      const size_type               numVectors,
+      LinAlgOpContext<memorySpace> *linAlgOpContext)
+    {
+      d_storage         = std::move(storage);
+      d_linAlgOpContext = linAlgOpContext;
+      d_vectorAttributes =
+        VectorAttributes(VectorAttributes::Distribution::SERIAL);
+      d_globalSize       = d_storage.size();
+      d_locallyOwnedSize = d_storage.size();
+      d_ghostSize        = 0;
+      d_localSize        = d_locallyOwnedSize + d_ghostSize;
+      d_numVectors       = numVectors;
+      d_mpiPatternP2P =
+        std::make_shared<const utils::mpi::MPIPatternP2P<memorySpace>>(
+          d_localSize);
+      d_mpiCommunicatorP2P = std::make_unique<
+        utils::mpi::MPICommunicatorP2P<ValueType, memorySpace>>(d_mpiPatternP2P,
+                                                                numVectors);
+    }
+
     //
-    // Constructor
+    // Constructor for \distributed MultiVector using an existing
+    // MPIPatternP2P object
     //
     template <typename ValueType, dftefe::utils::MemorySpace memorySpace>
     MultiVector<ValueType, memorySpace>::MultiVector(
-      std::unique_ptr<Storage> &    storage,
-      const global_size_type        globalSize,
-      const size_type               locallyOwnedSize,
-      const size_type               ghostSize,
+      std::shared_ptr<const utils::mpi::MPIPatternP2P<memorySpace>>
+                                    mpiPatternP2P,
+      LinAlgOpContext<memorySpace> *linAlgOpContext,
       const size_type               numVectors,
-      LinAlgOpContext<memorySpace> *linAlgOptContext)
-      : d_storage(storage)
-      , d_linAlgOptContext(linAlgOptContext)
-      , d_vectorAttributes(
-          VectorAttributes(VectorAttributes::Distribution::SERIAL))
-      , d_globalSize(globalSize)
-      , d_locallyOwnedSize(locallyOwnedSize)
-      , d_ghostSize(ghostSize)
-      , d_numVectors(numVectors)
+      const ValueType               initVal)
+      : d_mpiPatternP2P(mpiPatternP2P)
     {
-      d_localSize = locallyOwnedSize + ghostSize;
+      d_vectorAttributes =
+        VectorAttributes(VectorAttributes::Distribution::DISTRIBUTED);
+      d_globalSize       = d_mpiPatternP2P->nGlobalIndices();
+      d_locallyOwnedSize = d_mpiPatternP2P->localOwnedSize();
+      d_ghostSize        = d_mpiPatternP2P->localGhostSize();
+      d_localSize        = d_locallyOwnedSize + d_ghostSize;
+      d_numVectors       = numVectors;
+      d_storage =
+        std::make_unique<typename MultiVector<ValueType, memorySpace>::Storage>(
+          d_localSize * d_numVectors, initVal);
+      d_linAlgOpContext    = linAlgOpContext;
+      d_mpiCommunicatorP2P = std::make_unique<
+        utils::mpi::MPICommunicatorP2P<ValueType, memorySpace>>(mpiPatternP2P,
+                                                                numVectors);
     }
 
-    //
-    // Default Constructor
-    //
+    /**
+     * @brief Constructor for a \b distributed MultiVector with a predefined
+     * MultiVector::Storage (i.e., utils::MemoryStorage) and MPIPatternP2P.
+     * This constructor transfers the ownership of the input Storage to the
+     * MultiVector. This is useful when one does not want to allocate new
+     * memory and instead use memory allocated in the input MultiVector::Storage
+     * (i.e., utils::MemoryStorage).
+     */
     template <typename ValueType, dftefe::utils::MemorySpace memorySpace>
-    MultiVector<ValueType, memorySpace>::MultiVector()
-      : d_storage(nullptr)
-      , d_linAlgOptContext(nullptr)
-      , d_vectorAttributes(
-          VectorAttributes(VectorAttributes::Distribution::SERIAL))
-      , d_globalSize(0)
-      , d_locallyOwnedSize(0)
-      , d_ghostSize(0)
-      , d_localSize(0)
-      , d_numVectors(0)
-    {}
-
-    template <typename ValueType, dftefe::utils::MemorySpace memorySpace>
-    typename MultiVector<ValueType, memorySpace>::iterator
-    MultiVector<ValueType, memorySpace>::begin()
-    {
-      return d_storage->begin();
-    }
-
-    template <typename ValueType, dftefe::utils::MemorySpace memorySpace>
-    typename MultiVector<ValueType, memorySpace>::const_iterator
-    MultiVector<ValueType, memorySpace>::begin() const
-    {
-      return d_storage->begin();
-    }
-
-    template <typename ValueType, dftefe::utils::MemorySpace memorySpace>
-    typename MultiVector<ValueType, memorySpace>::iterator
-    MultiVector<ValueType, memorySpace>::end()
-    {
-      return d_storage->end();
-    }
-
-    template <typename ValueType, dftefe::utils::MemorySpace memorySpace>
-    typename MultiVector<ValueType, memorySpace>::const_iterator
-    MultiVector<ValueType, memorySpace>::end() const
-    {
-      return d_storage->end();
-    }
-
-    template <typename ValueType, dftefe::utils::MemorySpace memorySpace>
-    global_size_type
-    MultiVector<ValueType, memorySpace>::size() const
-    {
-      return d_globalSize;
-    }
-
-    template <typename ValueType, dftefe::utils::MemorySpace memorySpace>
-    size_type
-    MultiVector<ValueType, memorySpace>::locallyOwnedSize() const
-    {
-      return d_locallyOwnedSize;
-    }
-
-    template <typename ValueType, dftefe::utils::MemorySpace memorySpace>
-    size_type
-    MultiVector<ValueType, memorySpace>::ghostSize() const
-    {
-      return d_ghostSize;
-    }
-
-    template <typename ValueType, dftefe::utils::MemorySpace memorySpace>
-    size_type
-    MultiVector<ValueType, memorySpace>::localSize() const
-    {
-      return d_localSize;
-    }
-
-
-    template <typename ValueType, dftefe::utils::MemorySpace memorySpace>
-    size_type
-    MultiVector<ValueType, memorySpace>::numVectors() const
-    {
-      return d_numVectors;
-    }
-
-
-    template <typename ValueType, dftefe::utils::MemorySpace memorySpace>
-    ValueType *
-    MultiVector<ValueType, memorySpace>::data()
-    {
-      return d_storage->data();
-    }
-
-    template <typename ValueType, dftefe::utils::MemorySpace memorySpace>
-    const ValueType *
-    MultiVector<ValueType, memorySpace>::data() const
-    {
-      return d_storage->data();
-    }
-
-    template <typename ValueType, dftefe::utils::MemorySpace memorySpace>
-    MultiVector<ValueType, memorySpace> &
-    MultiVector<ValueType, memorySpace>::operator+=(
-      const MultiVector<ValueType, memorySpace> &rhs)
-    {
-      bool areCompatible =
-        d_vectorAttributes.areDistributionCompatible(rhs.getVectorAttributes());
-      utils::throwException<utils::LogicError>(
-        areCompatible,
-        "Trying to add incompatible MultiVectors. One is a serial MultiVector and the "
-        " other a distributed MultiVector.");
-      utils::throwException<utils::LengthError>(
-        rhs.size() == this->size() && rhs.localSize() == this->localSize() &&
-          rhs.numVectors() == this->numVectors(),
-        "Mismatch of sizes of the two MultiVectors that are being added.");
-      const size_type rhsStorageSize = (rhs.getValues()).size();
-      utils::throwException<utils::LengthError>(
-        d_storage->size() == rhsStorageSize,
-        "Mismatch of sizes of the underlying"
-        "storage of the two MultiVectors that are being added.");
-
-      blasLapack::axpby<ValueType, memorySpace>(
-        this->localSize() * this->numVectors(),
-        1.0,
-        this->data(),
-        1.0,
-        rhs.data(),
-        this->data(),
-        this->getLinAlgOptContext()->getBlasQueue());
-      return *this;
-    }
-
-    template <typename ValueType, dftefe::utils::MemorySpace memorySpace>
-    MultiVector<ValueType, memorySpace> &
-    MultiVector<ValueType, memorySpace>::operator-=(
-      const MultiVector<ValueType, memorySpace> &rhs)
-    {
-      bool areCompatible =
-        d_vectorAttributes.areDistributionCompatible(rhs.getVectorAttributes());
-      utils::throwException<utils::LogicError>(
-        areCompatible,
-        "Trying to add incompatible MultiVectors. "
-        "One is a serial vector and the other a distributed MultiVector.");
-      utils::throwException<utils::LengthError>(
-        rhs.size() == this->size() && rhs.localSize() == this->localSize() &&
-          rhs.numVectors() == this->numVectors(),
-        "Mismatch of sizes of the two MultiVectors that are being subtracted.");
-      const size_type rhsStorageSize = (rhs.getValues()).size();
-      utils::throwException<utils::LengthError>(
-        (d_storage->size() == rhsStorageSize),
-        "Mismatch of sizes of the underlying"
-        "storage of the two MultiVectors that are being subtracted.");
-      blasLapack::axpby<ValueType, memorySpace>(
-        this->localSize() * this->numVectors(),
-        1.0,
-        this->data(),
-        -1.0,
-        rhs.data(),
-        this->data(),
-        this->getLinAlgOptContext()->getBlasQueue());
-      return *this;
-    }
-
-    template <typename ValueType, dftefe::utils::MemorySpace memorySpace>
-    const typename MultiVector<ValueType, memorySpace>::Storage &
-    MultiVector<ValueType, memorySpace>::getValues() const
-    {
-      return *d_storage;
-    }
-
-    template <typename ValueType, dftefe::utils::MemorySpace memorySpace>
-    typename MultiVector<ValueType, memorySpace>::Storage &
-    MultiVector<ValueType, memorySpace>::getValues()
-    {
-      return *d_storage;
-    }
-
-    template <typename ValueType, dftefe::utils::MemorySpace memorySpace>
-    blasLapack::linAlgOptContext<memorySpace> *
-    MultiVector<ValueType, memorySpace>::getLinAlgOptContext() const
-    {
-      return d_linAlgOptContext;
-    }
-
-    template <typename ValueType, dftefe::utils::MemorySpace memorySpace>
-    template <dftefe::utils::MemorySpace memorySpace2>
-    void
-    MultiVector<ValueType, memorySpace>::setValues(
-      const typename MultiVector<ValueType, memorySpace2>::Storage &storage)
-    {
-      d_storage->copyFrom(storage);
-    }
-
-    template <typename ValueType, dftefe::utils::MemorySpace memorySpace>
-    void
-    MultiVector<ValueType, memorySpace>::setStorage(
+    MultiVector<ValueType, memorySpace>::MultiVector(
       std::unique_ptr<typename MultiVector<ValueType, memorySpace>::Storage>
-        &storage)
+        &storage,
+      std::shared_ptr<const utils::mpi::MPIPatternP2P<memorySpace>>
+                                    mpiPatternP2P,
+      LinAlgOpContext<memorySpace> *linAlgOpContext,
+      const size_type               numVectors)
+      : d_mpiPatternP2P(mpiPatternP2P)
     {
-      d_storage = std::move(storage);
+      d_storage         = std::move(storage);
+      d_linAlgOpContext = std::move(linAlgOpContext);
+      d_vectorAttributes =
+        VectorAttributes(VectorAttributes::Distribution::DISTRIBUTED);
+      d_globalSize         = d_mpiPatternP2P->nGlobalIndices();
+      d_locallyOwnedSize   = d_mpiPatternP2P->localOwnedSize();
+      d_ghostSize          = d_mpiPatternP2P->localGhostSize();
+      d_localSize          = d_locallyOwnedSize + d_ghostSize;
+      d_numVectors         = numVectors;
+      d_mpiCommunicatorP2P = std::make_unique<
+        utils::mpi::MPICommunicatorP2P<ValueType, memorySpace>>(mpiPatternP2P,
+                                                                numVectors);
+    }
+
+    /**
+     * @brief Constructor for a \distributed MultiVector based on locally
+     * owned and ghost indices.
+     * @note This way of construction is expensive. One should use the other
+     * constructor based on an input MPIPatternP2P as far as possible.
+     */
+    template <typename ValueType, dftefe::utils::MemorySpace memorySpace>
+    MultiVector<ValueType, memorySpace>::MultiVector(
+      const std::pair<global_size_type, global_size_type> locallyOwnedRange,
+      const std::vector<global_size_type> &               ghostIndices,
+      const utils::mpi::MPIComm &                         mpiComm,
+      LinAlgOpContext<memorySpace> *                      linAlgOpContext,
+      const size_type                                     numVectors,
+      const ValueType                                     initVal)
+    {
+      //
+      // TODO Move the warning message to a Logger class
+      //
+      int mpiRank;
+      int err = utils::mpi::MPICommRank(mpiComm, &mpiRank);
+      std::pair<bool, std::string> errIsSuccessAndMsg =
+        utils::mpi::MPIErrIsSuccessAndMsg(err);
+      utils::throwException(errIsSuccessAndMsg.first,
+                            errIsSuccessAndMsg.second);
+      std::string msg;
+      if (mpiRank == 0)
+        {
+          msg =
+            "WARNING: Constructing a distributed MultiVector using locally owned "
+            "range and ghost indices is expensive. As far as possible, one should "
+            " use the constructor based on an input MPIPatternP2P.";
+          std::cout << msg << std::endl;
+        }
+      ////////////
+
+      d_mpiPatternP2P =
+        std::make_shared<const utils::mpi::MPIPatternP2P<memorySpace>>(
+          locallyOwnedRange, ghostIndices, mpiComm);
+
+      d_mpiCommunicatorP2P = std::make_unique<
+        const utils::mpi::MPICommunicatorP2P<ValueType, memorySpace>>(
+        d_mpiPatternP2P, numVectors);
+
+      d_vectorAttributes = VectorAttributes::Distribution::DISTRIBUTED;
+      d_globalSize       = d_mpiPatternP2P->nGlobalIndices();
+      d_locallyOwnedSize = d_mpiPatternP2P->localOwnedSize();
+      d_ghostSize        = d_mpiPatternP2P->localGhostSize();
+      d_localSize        = d_locallyOwnedSize + d_ghostSize;
+      d_numVectors       = numVectors;
+      d_storage =
+        std::make_unique<typename MultiVector<ValueType, memorySpace>::Storage>(
+          d_localSize * numVectors, initVal);
+      d_linAlgOpContext = linAlgOpContext;
+    }
+
+    /**
+     * @brief Constructor for a special case of \b distributed MultiVector where none
+     * none of the processors have any ghost indices.
+     * @note This way of construction is expensive. One should use the other
+     * constructor based on an input MPIPatternP2P as far as possible.
+     */
+    template <typename ValueType, dftefe::utils::MemorySpace memorySpace>
+    MultiVector<ValueType, memorySpace>::MultiVector(
+      const std::pair<global_size_type, global_size_type> locallyOwnedRange,
+      const utils::mpi::MPIComm &                         mpiComm,
+      LinAlgOpContext<memorySpace> *                      linAlgOpContext,
+      const size_type                                     numVectors,
+      const ValueType                                     initVal)
+    {
+      //
+      // TODO Move the warning message to a Logger class
+      //
+      int mpiRank;
+      int err = utils::mpi::MPICommRank(mpiComm, &mpiRank);
+      std::pair<bool, std::string> errIsSuccessAndMsg =
+        utils::mpi::MPIErrIsSuccessAndMsg(err);
+      utils::throwException(errIsSuccessAndMsg.first,
+                            errIsSuccessAndMsg.second);
+      std::string msg;
+      if (mpiRank == 0)
+        {
+          msg =
+            "WARNING: Constructing a distributed MultiVector using only locally owned "
+            "range is expensive. As far as possible, one should "
+            " use the constructor based on an input MPIPatternP2P.";
+          std::cout << msg << std::endl;
+        }
+      ////////////
+      std::vector<dftefe::global_size_type> ghostIndices;
+      ghostIndices.resize(0);
+      d_mpiPatternP2P =
+        std::make_shared<const utils::mpi::MPIPatternP2P<memorySpace>>(
+          locallyOwnedRange, ghostIndices, mpiComm);
+
+      d_mpiCommunicatorP2P = std::make_unique<
+        const utils::mpi::MPICommunicatorP2P<ValueType, memorySpace>>(
+        d_mpiPatternP2P, numVectors);
+
+      d_vectorAttributes = VectorAttributes::Distribution::DISTRIBUTED;
+      d_globalSize       = d_mpiPatternP2P->nGlobalIndices();
+      d_locallyOwnedSize = d_mpiPatternP2P->localOwnedSize();
+      d_ghostSize        = d_mpiPatternP2P->localGhostSize();
+      d_localSize        = d_locallyOwnedSize + d_ghostSize;
+      d_numVectors       = numVectors;
+      d_storage =
+        std::make_unique<typename MultiVector<ValueType, memorySpace>::Storage>(
+          d_localSize * numVectors, initVal);
+      d_linAlgOpContext = linAlgOpContext;
+    }
+
+
+    /**
+     * @brief Constructor for a \b distributed MultiVector based on total number of global indices.
+     * The resulting Vector will not contain any ghost indices on any of the
+     * processors. Internally, the vector is divided to ensure as much
+     * equitable distribution across all the processors much as possible.
+     * @note This way of construction is expensive. One should use the other
+     * constructor based on an input MPIPatternP2P as far as possible.
+     * Further, the decomposition is not compatible with other ways of
+     * distributed MultiVector construction.
+     */
+    template <typename ValueType, dftefe::utils::MemorySpace memorySpace>
+    MultiVector<ValueType, memorySpace>::MultiVector(
+      const global_size_type        globalSize,
+      const utils::mpi::MPIComm &   mpiComm,
+      LinAlgOpContext<memorySpace> *linAlgOpContext,
+      const size_type               numVectors,
+      const ValueType               initVal)
+    {
+      std::vector<dftefe::global_size_type> ghostIndices;
+      ghostIndices.resize(0);
+
+      std::pair<global_size_type, global_size_type> locallyOwnedRange;
+
+      //
+      // TODO Move the warning message to a Logger class
+      //
+      int mpiRank;
+      int err = utils::mpi::MPICommRank(mpiComm, &mpiRank);
+      std::pair<bool, std::string> errIsSuccessAndMsg =
+        utils::mpi::MPIErrIsSuccessAndMsg(err);
+      utils::throwException(errIsSuccessAndMsg.first,
+                            errIsSuccessAndMsg.second);
+      std::string msg;
+      if (mpiRank == 0)
+        {
+          msg =
+            "WARNING: Constructing a MultiVector using total number of indices across all processors "
+            "is expensive. As far as possible, one should "
+            " use the constructor based on an input MPIPatternP2P.";
+          std::cout << msg << std::endl;
+        }
+
+      int mpiProcess;
+      int errProc        = utils::mpi::MPICommSize(mpiComm, &mpiProcess);
+      errIsSuccessAndMsg = utils::mpi::MPIErrIsSuccessAndMsg(errProc);
+      utils::throwException(errIsSuccessAndMsg.first,
+                            errIsSuccessAndMsg.second);
+
+      dftefe::global_size_type locallyOwnedSize = globalSize / mpiProcess;
+      if (mpiRank < globalSize % mpiProcess)
+        locallyOwnedSize++;
+
+      dftefe::global_size_type startIndex = mpiRank * (globalSize / mpiProcess);
+      if (mpiRank < globalSize % mpiProcess)
+        startIndex += mpiRank;
+      else
+        startIndex += globalSize % mpiProcess;
+
+      locallyOwnedRange.first  = startIndex;
+      locallyOwnedRange.second = startIndex + locallyOwnedSize;
+
+
+      ////////////
+
+      d_mpiPatternP2P =
+        std::make_shared<const utils::mpi::MPIPatternP2P<memorySpace>>(
+          locallyOwnedRange, ghostIndices, mpiComm);
+
+      d_mpiCommunicatorP2P = std::make_unique<
+        const utils::mpi::MPICommunicatorP2P<ValueType, memorySpace>>(
+        d_mpiPatternP2P, numVectors);
+
+      d_vectorAttributes = VectorAttributes::Distribution::DISTRIBUTED;
+      d_globalSize       = d_mpiPatternP2P->nGlobalIndices();
+      d_locallyOwnedSize = d_mpiPatternP2P->localOwnedSize();
+      d_ghostSize        = d_mpiPatternP2P->localGhostSize();
+      d_localSize        = d_locallyOwnedSize + d_ghostSize;
+      d_numVectors       = numVectors;
+      d_storage =
+        std::make_unique<typename MultiVector<ValueType, memorySpace>::Storage>(
+          d_localSize * numVectors, initVal);
+      d_linAlgOpContext = linAlgOpContext;
+    }
+
+
+
+    //
+    // Copy Constructor
+    //
+    template <typename ValueType, dftefe::utils::MemorySpace memorySpace>
+    MultiVector<ValueType, memorySpace>::MultiVector(const MultiVector &u)
+    {
+      d_storage =
+        std::make_unique<typename MultiVector<ValueType, memorySpace>::Storage>(
+          (u.d_storage)->size());
+      d_mpiCommunicatorP2P = std::make_unique<
+        utils::mpi::MPICommunicatorP2P<ValueType, memorySpace>>(
+        u.d_mpiPatternP2P, u.d_numVectors);
+      d_linAlgOpContext  = u.d_linAlgOpContext;
+      *d_storage         = *(u.d_storage);
+      d_vectorAttributes = u.d_vectorAttributes;
+      d_localSize        = u.d_localSize;
+      d_locallyOwnedSize = u.d_locallyOwnedSize;
+      d_ghostSize        = u.d_ghostSize;
+      d_globalSize       = u.d_globalSize;
+      d_numVectors       = u.d_numVectors;
+      d_mpiPatternP2P    = u.d_mpiPatternP2P;
+    }
+
+    //
+    // Copy Constructor with reinitialization
+    //
+    template <typename ValueType, dftefe::utils::MemorySpace memorySpace>
+    MultiVector<ValueType, memorySpace>::MultiVector(const MultiVector &u,
+                                                     const ValueType    initVal)
+    {
+      d_storage =
+        std::make_unique<typename MultiVector<ValueType, memorySpace>::Storage>(
+          (u.d_storage)->size(), initVal);
+      d_mpiCommunicatorP2P = std::make_unique<
+        utils::mpi::MPICommunicatorP2P<ValueType, memorySpace>>(
+        u.d_mpiPatternP2P, u.d_numVectors);
+      d_linAlgOpContext  = u.d_linAlgOpContext;
+      d_vectorAttributes = u.d_vectorAttributes;
+      d_localSize        = u.d_localSize;
+      d_locallyOwnedSize = u.d_locallyOwnedSize;
+      d_ghostSize        = u.d_ghostSize;
+      d_globalSize       = u.d_globalSize;
+      d_numVectors       = u.d_numVectors;
+      d_mpiPatternP2P    = u.d_mpiPatternP2P;
+    }
+
+    //
+    // Move Constructor
+    //
+    template <typename ValueType, dftefe::utils::MemorySpace memorySpace>
+    MultiVector<ValueType, memorySpace>::MultiVector(MultiVector &&u) noexcept
+    {
+      d_storage            = std::move(u.d_storage);
+      d_linAlgOpContext    = std::move(u.d_linAlgOpContext);
+      d_vectorAttributes   = std::move(u.d_vectorAttributes);
+      d_localSize          = std::move(u.d_localSize);
+      d_locallyOwnedSize   = std::move(u.d_locallyOwnedSize);
+      d_ghostSize          = std::move(u.d_ghostSize);
+      d_globalSize         = std::move(u.d_globalSize);
+      d_numVectors         = std::move(u.d_numVectors);
+      d_mpiCommunicatorP2P = std::move(u.d_mpiCommunicatorP2P);
+      d_mpiPatternP2P      = std::move(u.d_mpiPatternP2P);
+    }
+
+    //
+    // Copy Assignment
+    //
+    template <typename ValueType, dftefe::utils::MemorySpace memorySpace>
+    MultiVector<ValueType, memorySpace> &
+    MultiVector<ValueType, memorySpace>::operator=(const MultiVector &u)
+    {
+      d_storage =
+        std::make_unique<typename MultiVector<ValueType, memorySpace>::Storage>(
+          (u.d_storage)->size());
+      *d_storage           = *(u.d_storage);
+      d_mpiCommunicatorP2P = std::make_unique<
+        utils::mpi::MPICommunicatorP2P<ValueType, memorySpace>>(
+        u.d_mpiPatternP2P, u.d_numVectors);
+      d_linAlgOpContext  = u.d_linAlgOpContext;
+      d_vectorAttributes = u.d_vectorAttributes;
+      d_localSize        = u.d_localSize;
+      d_locallyOwnedSize = u.d_locallyOwnedSize;
+      d_ghostSize        = u.d_ghostSize;
+      d_globalSize       = u.d_globalSize;
+      d_numVectors       = u.d_numVectors;
+      d_mpiPatternP2P    = u.d_mpiPatternP2P;
+      return *this;
+    }
+
+    //
+    // Move Assignment
+    //
+    template <typename ValueType, dftefe::utils::MemorySpace memorySpace>
+    MultiVector<ValueType, memorySpace> &
+    MultiVector<ValueType, memorySpace>::operator=(MultiVector &&u)
+    {
+      d_storage            = std::move(u.d_storage);
+      d_linAlgOpContext    = std::move(u.d_linAlgOpContext);
+      d_vectorAttributes   = std::move(u.d_vectorAttributes);
+      d_localSize          = std::move(u.d_localSize);
+      d_locallyOwnedSize   = std::move(u.d_locallyOwnedSize);
+      d_ghostSize          = std::move(u.d_ghostSize);
+      d_globalSize         = std::move(u.d_globalSize);
+      d_numVectors         = std::move(u.d_numVectors);
+      d_mpiCommunicatorP2P = std::move(u.d_mpiCommunicatorP2P);
+      d_mpiPatternP2P      = std::move(u.d_mpiPatternP2P);
+      return *this;
     }
 
     template <typename ValueType, dftefe::utils::MemorySpace memorySpace>
-    const VectorAttributes &
-    MultiVector<ValueType, memorySpace>::getVectorAttributes() const
+    std::vector<double>
+    MultiVector<ValueType, memorySpace>::l2Norms() const
     {
-      return d_vectorAttributes;
+      const std::vector<double> l2NormsLocallyOwned =
+        blasLapack::nrms2MultiVector<ValueType, memorySpace>(
+          this->locallyOwnedSize(),
+          this->numVectors(),
+          this->data(),
+          *d_linAlgOpContext);
+
+      std::vector<double> l2NormsLocallyOwnedSquare(d_numVectors, 0.0);
+      for (size_type i = 0; i < d_numVectors; ++i)
+        l2NormsLocallyOwnedSquare[i] =
+          l2NormsLocallyOwned[i] * l2NormsLocallyOwned[i];
+
+      std::vector<double> returnValues(d_numVectors, 0.0);
+      utils::mpi::MPIAllreduce<memorySpace>(&l2NormsLocallyOwnedSquare,
+                                            &returnValues[0],
+                                            d_numVectors,
+                                            utils::mpi::MPIDouble,
+                                            utils::mpi::MPIMax,
+                                            d_mpiPatternP2P->mpiCommunicator());
+      for (size_type i = 0; i < d_numVectors; ++i)
+        returnValues[i] = std::sqrt(returnValues[i]);
+      return returnValues;
     }
 
-    //
-    // Helper functions
-    //
+    template <typename ValueType, dftefe::utils::MemorySpace memorySpace>
+    std::vector<double>
+    MultiVector<ValueType, memorySpace>::lInfNorms() const
+    {
+      const std::vector<double> lInfNormsLocallyOwned =
+        blasLapack::amaxsMultiVector<ValueType, memorySpace>(
+          this->locallyOwnedSize(),
+          this->numVectors(),
+          this->data(),
+          *d_linAlgOpContext);
+
+      std::vector<double> returnValues(d_numVectors, 0.0);
+      utils::mpi::MPIAllreduce<memorySpace>(&lInfNormsLocallyOwned,
+                                            &returnValues[0],
+                                            d_numVectors,
+                                            utils::mpi::MPIDouble,
+                                            utils::mpi::MPIMax,
+                                            d_mpiPatternP2P->mpiCommunicator());
+      return returnValues;
+    }
+
     template <typename ValueType, dftefe::utils::MemorySpace memorySpace>
     void
-    add(ValueType                                  a,
-        const MultiVector<ValueType, memorySpace> &u,
-        ValueType                                  b,
-        const MultiVector<ValueType, memorySpace> &v,
-        MultiVector<ValueType, memorySpace> &      w)
+    MultiVector<ValueType, memorySpace>::updateGhostValues(
+      const size_type communicationChannel /*= 0*/)
     {
-      const VectorAttributes &uVectorAttributes = u.getVectorAttributes();
-      const VectorAttributes &vVectorAttributes = v.getVectorAttributes();
-      const VectorAttributes &wVectorAttributes = w.getVectorAttributes();
-      bool                    areCompatible =
-        uVectorAttributes.areDistributionCompatible(vVectorAttributes);
-      utils::throwException(
-        areCompatible,
-        "Trying to add incompatible MultiVectors. One is a SerialVector and the other a DistributedVector");
-      areCompatible =
-        vVectorAttributes.areDistributionCompatible(wVectorAttributes);
-      utils::throwException(
-        areCompatible,
-        "Trying to add incompatible MultiVectors. One is a serialVector and the other a DistributedVector.");
-      utils::throwException<utils::LengthError>(
-        ((u.size() == v.size()) && (v.size() == w.size()) &&
-         (u.localSize() == v.localSize()) && (v.localSize() == w.localSize()) &&
-         (u.numVectors() == v.numVectors()) &&
-         (v.numVectors() == w.numVectors())),
-        "Mismatch of sizes of the MultiVectors that are added.");
-      const size_type uStorageSize = (u.getValues()).size();
-      const size_type vStorageSize = (v.getValues()).size();
-      const size_type wStorageSize = (w.getValues()).size();
-      utils::throwException<utils::LengthError>(
-        (uStorageSize == vStorageSize) && (vStorageSize == wStorageSize),
-        "Mismatch of sizes of the underlying storages"
-        "of the MultiVectors that are added.");
-
-      blasLapack::axpby<ValueType, memorySpace>(
-        u.localSize() * u.numVectors(),
-        a,
-        u.data(),
-        b,
-        v.data(),
-        w.data(),
-        w.getLinAlgOptContext()->getBlasQueue());
+      d_mpiCommunicatorP2P->updateGhostValues(*d_storage, communicationChannel);
     }
-  } // namespace linearAlgebra
+
+    template <typename ValueType, dftefe::utils::MemorySpace memorySpace>
+    void
+    MultiVector<ValueType, memorySpace>::accumulateAddLocallyOwned(
+      const size_type communicationChannel /*= 0*/)
+    {
+      d_mpiCommunicatorP2P->accumulateAddLocallyOwned(*d_storage,
+                                                      communicationChannel);
+    }
+
+    template <typename ValueType, dftefe::utils::MemorySpace memorySpace>
+    void
+    MultiVector<ValueType, memorySpace>::updateGhostValuesBegin(
+      const size_type communicationChannel /*= 0*/)
+    {
+      d_mpiCommunicatorP2P->updateGhostValuesBegin(*d_storage,
+                                                   communicationChannel);
+    }
+
+    template <typename ValueType, dftefe::utils::MemorySpace memorySpace>
+    void
+    MultiVector<ValueType, memorySpace>::updateGhostValuesEnd()
+    {
+      d_mpiCommunicatorP2P->updateGhostValuesEnd(*d_storage);
+    }
+
+    template <typename ValueType, dftefe::utils::MemorySpace memorySpace>
+    void
+    MultiVector<ValueType, memorySpace>::accumulateAddLocallyOwnedBegin(
+      const size_type communicationChannel /*= 0*/)
+    {
+      d_mpiCommunicatorP2P->accumulateAddLocallyOwnedBegin(
+        *d_storage, communicationChannel);
+    }
+
+    template <typename ValueType, dftefe::utils::MemorySpace memorySpace>
+    void
+    MultiVector<ValueType, memorySpace>::accumulateAddLocallyOwnedEnd()
+    {
+      d_mpiCommunicatorP2P->accumulateAddLocallyOwnedEnd(*d_storage);
+    }
+  } // end of namespace linearAlgebra
 } // namespace dftefe
